# projects/gin-advanced/docs/swagger.yaml
basePath: /
definitions:
...
  api.AddPermissionRequest:
    properties:
      permissionID:
        type: integer
      roleID:
        type: integer
    required:
    - permissionID
    - roleID
    type: object
  api.AssignRoleRequest:
    properties:
      roleID:
        type: integer
      userID:
        type: integer
    required:
    - roleID
    - userID
    type: object
...
  api.SetRolePermissionsRequest:
    properties:
      permission_ids:
        items:
          type: integer
        type: array
      role_id:
        type: integer
    required:
    - role_id
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Permission:
    properties:
      created_at:
        type: integer
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        description: 权限描述
        type: string
      id:
        type: integer
      method:
        description: 请求方法 GET/POST/PUT/DELETE
        type: string
      name:
        description: 权限名称，如 "user:create"
        type: string
      path:
        description: 请求路径
        type: string
      updated_at:
        type: integer
    type: object
  models.Role:
    properties:
      created_at:
        type: integer
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        description: 角色描述
        type: string
      id:
        type: integer
      name:
        description: 角色名称
        type: string
      permissions:
        description: 权限关联
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      updated_at:
        type: integer
    type: object
...
  utils.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Gin-Advanced RESTful API 文档
  title: Gin-Advanced API
  version: "1.0"
paths:
...
  /admin/permissions:
    post:
      consumes:
      - application/json
      description: 创建一个新的权限
      parameters:
      - description: 权限信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Permission'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Permission'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 创建权限
      tags:
      - 角色管理
  /admin/permissions/{id}:
    delete:
      description: 删除指定权限
      parameters:
      - description: 权限ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 删除权限
      tags:
      - 角色管理
    put:
      consumes:
      - application/json
      description: 更新指定权限的信息
      parameters:
      - description: 权限信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Permission'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Permission'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 更新权限
      tags:
      - 角色管理
  /admin/roles:
    get:
      description: 获取所有角色的列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Role'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 获取角色列表
      tags:
      - 角色管理
    post:
      consumes:
      - application/json
      description: 创建一个新的角色
      parameters:
      - description: 角色信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Role'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Role'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 创建角色
      tags:
      - 角色管理
  /admin/roles/assign:
    post:
      consumes:
      - application/json
      description: 为指定用户分配一个角色
      parameters:
      - description: 分配信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.AssignRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 给用户分配角色
      tags:
      - 角色管理
  /admin/roles/permission:
    get:
      description: 获取系统中所有可用的权限列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Permission'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 获取所有权限
      tags:
      - 角色管理
    post:
      consumes:
      - application/json
      description: 为指定角色添加一个权限
      parameters:
      - description: 权限信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.AddPermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 给角色添加权限
      tags:
      - 角色管理
    put:
      consumes:
      - application/json
      description: 替换指定角色的所有权限
      parameters:
      - description: 权限列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.SetRolePermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 批量设置角色权限
      tags:
      - 角色管理
  /admin/seed:
    get:
      description: 创建默认的admin角色和权限，并分配给admin用户
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 初始化默认角色和权限
      tags:
      - 角色管理
...
  /upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传通用文件到服务器或OSS
      parameters:
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 上传文件
      tags:
      - 文件上传
  /upload/avatar:
    post:
      consumes:
      - multipart/form-data
      description: 上传用户头像图片
      parameters:
      - description: 头像文件
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 上传头像
      tags:
      - 文件上传
...
  /upload/image:
    post:
      consumes:
      - multipart/form-data
      description: 上传图片到服务器或OSS
      parameters:
      - description: 图片文件
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 上传图片
      tags:
      - 文件上传
  /upload/multiple:
    post:
      consumes:
      - multipart/form-data
      description: 一次上传多个文件
      parameters:
      - description: 文件列表
        in: formData
        name: files
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 批量上传文件
      tags:
      - 文件上传
...
securityDefinitions:
  BearerAuth:
    in: header
    name: token
    type: apiKey
swagger: "2.0"